Projects/Web Development/Ticket Show/ticket_show/code/backend/requirements.txt
Flask==2.2.0
jinja2
Flask-SQLAlchemy==3.0.3
flask-security-too
email_validator
bcrypt
flask_restful
flask-cors
pytz
redis
celery[redis]
flask_sse
weasyprint==52.5
Flask-Caching